% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{availableDataSets}
\alias{availableDataSets}
\title{Available data sets}
\usage{
availableDataSets(dataSetDir = "DataSets", internalDir = TRUE)
}
\arguments{
\item{dataSetDir}{directory containing local data set store. If \code{internalDir = TRUE} the full directory path would be relative to the package installation location.}

\item{internalDir}{stored the data set internally to the package installation location}
}
\value{
A tibble containing available data set information.
}
\description{
Displays available data sets.
}
\examples{
\dontrun{
availableDataSets()
}
}
