% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{availableFiles}
\alias{availableFiles}
\title{Available files for a data set}
\usage{
availableFiles(technique, dataSet, dataSetDir = "DataSets", internalDir = TRUE)
}
\arguments{
\item{technique}{metabolomic technique name}

\item{dataSet}{data set name}

\item{dataSetDir}{directory containing local data set store. If \code{internalDir = TRUE} the full directory path would be relative to the package installation location.}

\item{internalDir}{stored the data set internally to the package installation location}
}
\value{
A tibble containing available file information.
}
\description{
Display the available files for a given data set.
}
\examples{
\dontrun{
availableFiles('FIE-HRMS','BdistachyonTechnical')
}
}
