% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{dataSets}
\alias{dataSets}
\title{Available data sets for a given technique}
\usage{
dataSets(technique)
}
\arguments{
\item{technique}{metabolomic technique name}
}
\value{
A character vector of available data sets.
}
\description{
Return a vector of available data sets for a given metabolomic technique.
}
\examples{
\donttest{
dataSets(techniques()[1])
}
}
