% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{description}
\alias{description}
\title{description}
\usage{
description(technique, dataSet)
}
\arguments{
\item{technique}{metabolomic technique name}

\item{dataSet}{data set name}
}
\description{
Return list containing experiment description elements
}
\examples{
experimentDescription <- description(
   techniques()[1],
   dataSets(techniques()[1])[1])
}
