% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadDataSet.R
\name{downloadDataSet}
\alias{downloadDataSet}
\title{Download a data set}
\usage{
downloadDataSet(
  technique,
  dataSet,
  dataSetDir = "DataSets",
  internalDir = TRUE,
  ask = TRUE
)
}
\arguments{
\item{technique}{metabolomic technique name}

\item{dataSet}{data set name}

\item{dataSetDir}{directory to store local data sets. If \code{internalDir = TRUE} this is a directory relative to the library location.}

\item{internalDir}{Logical, should the directory for storing local data sets be internal to the package location.}

\item{ask}{ask before downloading}
}
\description{
Download an example data set.
}
