% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runinfo.R
\name{runinfo}
\alias{runinfo}
\title{runinfo}
\usage{
runinfo(technique, dataSet)
}
\arguments{
\item{technique}{metabolomic technique name}

\item{dataSet}{data set name}
}
\description{
Return tibble of run information for a given data set of a given metabolomics technique.
}
\examples{
info <- runinfo(
   techniques()[1],
   dataSets(techniques()[1])[1])

head(info)
}
