% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{techniques}
\alias{techniques}
\title{Available data set techniques}
\usage{
techniques()
}
\value{
A character vector of available techniques.
}
\description{
Return a vector of available metabolomic techniques.
}
\examples{
techniques()
}
