% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableDataSets.R
\name{availableDataSets}
\alias{availableDataSets}
\title{availableDataSets}
\usage{
availableDataSets()
}
\description{
Displays file paths for available data sets.
}
\examples{
availableDataSets()

## Can be run to access individual file paths
list.files(system.file('DataSets/FIE-HRMS/BdistachyonEcotypes',
                        package = 'metaboData'),
           full.names = TRUE)
}
