% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{dataSets}
\alias{dataSets}
\title{dataSets}
\usage{
dataSets(technique)
}
\arguments{
\item{technique}{metabolomic technique name}
}
\description{
Return a vector of available data sets for a given metabolomic technique.
}
\examples{
dataSets(techniques()[1])
}
