% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{filePaths}
\alias{filePaths}
\title{filePaths}
\usage{
filePaths(technique, dataSet)
}
\arguments{
\item{technique}{metabolomic technique name}

\item{dataSet}{data set name}
}
\description{
Return a vector of file paths for a given data set of a given metabolomic technique.
}
\examples{
files <- filePaths(
   techniques()[1],
   dataSets(techniques()[1])[1])

head(files)
}
